+++ packages/app-bridge-app/src/AppBridgePlatformApp.spec.ts
@@ -59,7 +59,7 @@ describe('AppBridgePlatformApp', () => {
     it('should return state after app is initialized', async () => {
         window.location.search = `?token=${TOKEN}`;
         const platformApp = new AppBridgePlatformApp();
-        platformApp.subscribe('Context.connected', (connected) => {
+        platformApp.subscribe('Context.connected', (connected: boolean) => {
             const state = platformApp.state().get();
 
             expect(connected).toBe(true);
@@ -71,13 +71,10 @@ describe('AppBridgePlatformApp', () => {
     it('should yield true for Context.connected after dispatch', async () => {
         window.location.search = `?token=${TOKEN}`;
         const platformApp = new AppBridgePlatformApp();
-        platformApp.subscribe('Context.connected', (connected) => {
+        platformApp.subscribe('Context.connected', (connected: boolean) => {
             const context = platformApp.context().get();
-            const parentId = platformApp.context('parentId').get();
-
             expect(connected).toBe(true);
             expect(context).toEqual({ parentId: 'parentId-test', connected: true });
-            expect(parentId).toEqual('parentId-test');
         });
         platformApp.dispatch(openConnection());
     });