+++ packages/guideline-blocks-settings/src/hooks/useDndSensors.ts
@@ -1,6 +1,7 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 import { KeyboardSensor, PointerSensor, useSensor, useSensors } from '@dnd-kit/core';
+import { useMemo } from 'react';
 
 import { customCoordinatesGetterFactory } from '../helpers/customCoordinatesGetterFactory';
 
@@ -11,14 +12,15 @@ const keyboardCodes = {
 };
 
 export const useDndSensors = (columnGap = 0, rowGap = 0) => {
-    const customCoordinatesGetter = customCoordinatesGetterFactory(columnGap, rowGap);
-    const sensors = useSensors(
-        useSensor(PointerSensor),
-        useSensor(KeyboardSensor, {
+    const keyboardSensorOptions = useMemo(() => {
+        const customCoordinatesGetter = customCoordinatesGetterFactory(columnGap, rowGap);
+        return {
             coordinateGetter: customCoordinatesGetter,
             keyboardCodes,
-        }),
-    );
+        };
+    }, [columnGap, rowGap]);
+
+    const sensors = useSensors(useSensor(PointerSensor), useSensor(KeyboardSensor, keyboardSensorOptions));
 
     return sensors;
 };