+++ packages/sidebar-settings/src/blocks/fontInput.ts
@@ -190,6 +190,38 @@ export type FontValue = {
     highlightBorderRadiusChoice?: BorderRadius;
 
     /**
+     * Highlight border settings
+     */
+    highlightBorderEnabled?: boolean;
+    highlightBorderColor?: {
+        /**
+         * The red value of the color (0-255).
+         */
+        red: number;
+        /**
+         * The green value of the color (0-255).
+         */
+        green: number;
+        /**
+         * The blue value of the color (0-255).
+         */
+        blue: number;
+        /**
+         * The alpha value of the color (0-1).
+         */
+        alpha?: number;
+    } | null;
+    highlightBorderStyle?: 'solid' | 'dashed' | 'dotted';
+    highlightBorderWeight?: string;
+    highlightBorderWeightCustomEnabled?: boolean;
+    highlightBorderWeightCustom?: {
+        top: string;
+        right: string;
+        bottom: string;
+        left: string;
+    } | null;
+
+    /**
      * The rgba-color of the text on hover.
      */
     hoverColor?: {
@@ -288,6 +320,38 @@ export type FontValue = {
     hoverHighlightBorderRadiusChoice?: BorderRadius;
 
     /**
+     * Highlight border settings on hover
+     */
+    hoverHighlightBorderEnabled?: boolean;
+    hoverHighlightBorderColor?: {
+        /**
+         * The red value of the color (0-255).
+         */
+        red: number;
+        /**
+         * The green value of the color (0-255).
+         */
+        green: number;
+        /**
+         * The blue value of the color (0-255).
+         */
+        blue: number;
+        /**
+         * The alpha value of the color (0-1).
+         */
+        alpha?: number;
+    } | null;
+    hoverHighlightBorderStyle?: 'solid' | 'dashed' | 'dotted';
+    hoverHighlightBorderWeight?: string;
+    hoverHighlightBorderWeightCustomEnabled?: boolean;
+    hoverHighlightBorderWeightCustom?: {
+        top: string;
+        right: string;
+        bottom: string;
+        left: string;
+    } | null;
+
+    /**
      * The color of the text when active.
      */
     activeColor?: {
@@ -384,6 +448,38 @@ export type FontValue = {
     activeHighlightBorderRadiusCustomEnabled?: boolean;
     activeHighlightBorderRadiusCustom?: FontValueBorderRadiusType | null;
     activeHighlightBorderRadiusChoice?: BorderRadius;
+
+    /**
+     * Highlight border settings when active
+     */
+    activeHighlightBorderEnabled?: boolean;
+    activeHighlightBorderColor?: {
+        /**
+         * The red value of the color (0-255).
+         */
+        red: number;
+        /**
+         * The green value of the color (0-255).
+         */
+        green: number;
+        /**
+         * The blue value of the color (0-255).
+         */
+        blue: number;
+        /**
+         * The alpha value of the color (0-1).
+         */
+        alpha?: number;
+    } | null;
+    activeHighlightBorderStyle?: 'solid' | 'dashed' | 'dotted';
+    activeHighlightBorderWeight?: string;
+    activeHighlightBorderWeightCustomEnabled?: boolean;
+    activeHighlightBorderWeightCustom?: {
+        top: string;
+        right: string;
+        bottom: string;
+        left: string;
+    } | null;
 };
 
 export type FontInputBlock<AppBridge> = {